Adding forwarders to create_network
This commit adds a simple implementation for adding forwarders to the
network XML definition.

Forwarders are useful when the dnsmasq instance serving DNS and DHCP for
the network must forward to a non-default upstream DNS server. The
default is to use the DNS server configured in /etc/resolv.conf.

It can be useful to *not* use the DNS server in /etc/resolv.conf to
prevent DNS loops (which can happen if systemd-resolved is configured to
query the libvirt dnsmasq instance for some hostnames).